refactor: 의미가 명확하지 않은 패키지와 메서드를 명확하게 수정 및 구조 변경
