+++ docs/README.md
@@ -25,7 +25,10 @@
 ## 1 단계 피드백, 수정사항
 
 - [ ] Controller와 Converter는 앞에 문자도 겹치고 각 도메인별로 모두 있어서 헷갈릴 수 있을 것 같은데요, 패키지를 분리해 보면 어떨까요?
-- [ ] (컨트롤러) 보통 Info라는 단어는 명확한 의미를 드러내지 못하는 경우가 많아 지양하는 것이 좋은데요, 여기서도 마찬가지인 것 같아요.
+- [x] (컨트롤러) 보통 Info라는 단어는 명확한 의미를 드러내지 못하는 경우가 많아 지양하는 것이 좋은데요, 여기서도 마찬가지인 것 같아요.
+  - 코드에서 Controller가 요청을 받거나 응답을 주는 객체를 그대로 서비스에서 쓰지 않고 서비스와 컨트롤러 사이의 dto가 있습니다.
+  - toInfo와 dto.info의 의미는 컨트롤러와 서비스 사이에 정보를 주고받는 dto를 의미합니다.
+  - 하지만 제가 생각해도 이해가 쉽지 않다고 느꼈고 이부분을 조금 더 명확하게 toServiceRequest로 수정하고 dto.info를 dto.service로 수정했습니다.
 - [x] (Advice) IllegalArgumentException만 핸들링하면 될까?
   - Advice에서는 애플리케이션에서 내가 정할 수 있는 가능한 모든 에러를 핸들링합니다.
   - 하지만 예상하지 못하는 에러가 생길 수 있기 때문에 내가 처리하지 못한 에러도 핸들할 수 있도록 `Exception.class`도 추가했습니다.
