+++ src/main/java/wooteco/subway/controller/PathController.java
@@ -5,9 +5,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
-import wooteco.subway.dto.info.PathServiceRequest;
-import wooteco.subway.dto.response.PathResponse;
+import wooteco.subway.dto.controller.response.PathResponse;
+import wooteco.subway.dto.service.PathServiceRequest;
 import wooteco.subway.service.PathService;
 
 @RestController
@@ -23,7 +22,7 @@ public class PathController {
     @GetMapping
     public ResponseEntity<PathResponse> getShortestPath(@RequestParam Long source, @RequestParam Long target,
         @RequestParam Integer age) {
-        PathServiceRequest pathServiceRequest = PathConverter.toInfo(source, target, age);
+        PathServiceRequest pathServiceRequest = PathConverter.toServiceRequest(source, target, age);
         PathResponse pathResponse = PathConverter.toResponse(pathService.getShortestPath(pathServiceRequest));
         return ResponseEntity.ok(pathResponse);
     }
