+++ src/main/java/wooteco/subway/controller/SectionController.java
@@ -8,8 +8,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
-import wooteco.subway.dto.request.SectionRequest;
+import wooteco.subway.dto.controller.request.SectionRequest;
 import wooteco.subway.service.SectionService;
 
 @RestController
@@ -24,13 +23,13 @@ public class SectionController {
 
     @PostMapping
     public ResponseEntity<Void> createSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
-        sectionService.save(SectionConverter.toInfo(id, sectionRequest));
+        sectionService.save(SectionConverter.toServiceRequest(id, sectionRequest));
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping
     public ResponseEntity<Void> deleteSection(@PathVariable Long id, @RequestParam Long stationId) {
-        sectionService.delete(SectionConverter.toInfo(id, stationId));
+        sectionService.delete(SectionConverter.toServiceRequest(id, stationId));
         return ResponseEntity.ok().build();
     }
 }
