+++ src/main/java/wooteco/subway/controller/SectionConverter.java
@@ -1,17 +1,17 @@
 package wooteco.subway.controller;
 
-import wooteco.subway.dto.info.SectionCreateRequest;
-import wooteco.subway.dto.info.SectionDeleteRequest;
-import wooteco.subway.dto.request.SectionRequest;
+import wooteco.subway.dto.controller.request.SectionRequest;
+import wooteco.subway.dto.service.SectionCreateRequest;
+import wooteco.subway.dto.service.SectionDeleteRequest;
 
 public class SectionConverter {
 
-    static SectionCreateRequest toInfo(Long lineId, SectionRequest sectionRequest) {
+    static SectionCreateRequest toServiceRequest(Long lineId, SectionRequest sectionRequest) {
         return new SectionCreateRequest(lineId, sectionRequest.getUpStationId(), sectionRequest.getDownStationId(),
             sectionRequest.getDistance());
     }
 
-    static SectionDeleteRequest toInfo(Long lineId, Long stationId) {
+    static SectionDeleteRequest toServiceRequest(Long lineId, Long stationId) {
         return new SectionDeleteRequest(lineId, stationId);
     }
 }
