+++ src/main/java/wooteco/subway/controller/StationConverter.java
@@ -1,16 +1,12 @@
 package wooteco.subway.controller;
 
-import wooteco.subway.dto.info.StationDto;
-import wooteco.subway.dto.request.StationRequest;
-import wooteco.subway.dto.response.StationResponse;
+import wooteco.subway.dto.controller.request.StationRequest;
+import wooteco.subway.dto.controller.response.StationResponse;
+import wooteco.subway.dto.service.StationDto;
 
 public class StationConverter {
-    static StationDto toInfo(StationRequest stationRequest) {
+    static StationDto toServiceRequest(StationRequest stationRequest) {
         return new StationDto(stationRequest.getName());
-    }
-
-    static StationDto toInfo(Long id, StationRequest stationRequest) {
-        return new StationDto(id, stationRequest.getName());
     }
 
     static StationResponse toResponse(StationDto stationDto) {
