+++ src/main/java/wooteco/subway/dao/JdbcLineDao.java
@@ -3,15 +3,13 @@ package wooteco.subway.dao;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.util.List;
-
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-
 import wooteco.subway.domain.Line;
-import wooteco.subway.dto.LineEntity;
+import wooteco.subway.entity.LineEntity;
 
 @Repository
 public class JdbcLineDao implements LineDao {
