+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -1,7 +1,6 @@
 package wooteco.subway.dao;
 
 import java.util.List;
-
 import wooteco.subway.domain.Station;
 
 public interface StationDao {
