+++ src/main/java/wooteco/subway/domain/Path.java
@@ -1,7 +1,6 @@
 package wooteco.subway.domain;
 
 import java.util.List;
-
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
