+++ src/main/java/wooteco/subway/dto/controller/response/LineResponse.java
@@ -1,16 +1,21 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.controller.response;
 
-public class LineEntity {
+import java.util.List;
+
+public class LineResponse {
     private final long id;
     private final String name;
     private final String color;
     private final int extraFare;
+    private final List<StationResponse> stations;
 
-    public LineEntity(long id, String name, String color, int extraFare) {
+    public LineResponse(long id, String name, String color,
+        int extraFare, List<StationResponse> stations) {
         this.id = id;
         this.name = name;
         this.color = color;
         this.extraFare = extraFare;
+        this.stations = stations;
     }
 
     public long getId() {
@@ -27,5 +32,9 @@ public class LineEntity {
 
     public int getExtraFare() {
         return extraFare;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
     }
 }
