+++ src/main/java/wooteco/subway/dto/service/LineServiceResponse.java
@@ -1,21 +1,21 @@
-package wooteco.subway.dto.response;
+package wooteco.subway.dto.service;
 
 import java.util.List;
 
-public class LineResponse {
+public class LineServiceResponse {
     private final long id;
     private final String name;
     private final String color;
     private final int extraFare;
-    private final List<StationResponse> stations;
+    private final List<StationDto> stationDtos;
 
-    public LineResponse(long id, String name, String color,
-        int extraFare, List<StationResponse> stations) {
+    public LineServiceResponse(long id, String name, String color,
+        int extraFare, List<StationDto> stationDtos) {
         this.id = id;
         this.name = name;
         this.color = color;
         this.extraFare = extraFare;
-        this.stations = stations;
+        this.stationDtos = stationDtos;
     }
 
     public long getId() {
@@ -34,7 +34,7 @@ public class LineResponse {
         return extraFare;
     }
 
-    public List<StationResponse> getStations() {
-        return stations;
+    public List<StationDto> getStationInfos() {
+        return stationDtos;
     }
 }
