+++ src/main/java/wooteco/subway/dto/service/LineUpdateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.info;
+package wooteco.subway.dto.service;
 
 public class LineUpdateRequest {
     private long id;
