+++ src/main/java/wooteco/subway/dto/service/PathServiceRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.info;
+package wooteco.subway.dto.service;
 
 public class PathServiceRequest {
     private final long source;
