+++ src/main/java/wooteco/subway/dto/service/PathServiceResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.info;
+package wooteco.subway.dto.service;
 
 import java.util.List;
