+++ src/main/java/wooteco/subway/dto/service/SectionDeleteRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.info;
+package wooteco.subway.dto.service;
 
 public class SectionDeleteRequest {
     private final long lineId;
