+++ src/main/java/wooteco/subway/entity/LineEntity.java
@@ -1,21 +1,16 @@
-package wooteco.subway.dto.info;
+package wooteco.subway.entity;
 
-import java.util.List;
-
-public class LineServiceResponse {
+public class LineEntity {
     private final long id;
     private final String name;
     private final String color;
     private final int extraFare;
-    private final List<StationDto> stationDtos;
 
-    public LineServiceResponse(long id, String name, String color,
-        int extraFare, List<StationDto> stationDtos) {
+    public LineEntity(long id, String name, String color, int extraFare) {
         this.id = id;
         this.name = name;
         this.color = color;
         this.extraFare = extraFare;
-        this.stationDtos = stationDtos;
     }
 
     public long getId() {
@@ -32,9 +27,5 @@ public class LineServiceResponse {
 
     public int getExtraFare() {
         return extraFare;
-    }
-
-    public List<StationDto> getStationInfos() {
-        return stationDtos;
     }
 }
