+++ src/main/java/wooteco/subway/strategy/FareStrategy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.strategy;
 
 public interface FareStrategy {
     int calculate(int distance);
