+++ src/test/java/wooteco/subway/dao/JdbcLineDaoTest.java
@@ -1,6 +1,6 @@
 package wooteco.subway.dao;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -8,9 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-
 import wooteco.subway.domain.Line;
-import wooteco.subway.dto.LineEntity;
+import wooteco.subway.entity.LineEntity;
 
 @JdbcTest
 public class JdbcLineDaoTest {
