+++ src/test/java/wooteco/subway/domain/LineTest.java
@@ -1,6 +1,6 @@
 package wooteco.subway.domain;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
