+++ src/test/java/wooteco/subway/service/FakeLineDao.java
@@ -5,12 +5,10 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
 import org.springframework.util.ReflectionUtils;
-
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.Line;
-import wooteco.subway.dto.LineEntity;
+import wooteco.subway.entity.LineEntity;
 
 public class FakeLineDao implements LineDao {
     private Long seq = 0L;
