+++ src/test/java/wooteco/subway/service/FakeStationDao.java
@@ -5,9 +5,7 @@ import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-
 import org.springframework.util.ReflectionUtils;
-
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
