+++ src/test/java/wooteco/subway/service/StationServiceTest.java
@@ -1,16 +1,16 @@
 package wooteco.subway.service;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.NoSuchElementException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
-import wooteco.subway.dto.info.StationDto;
+import wooteco.subway.dto.service.StationDto;
 
 public class StationServiceTest {
     private StationService stationService;
